Sprint6/Lab2/flask-app/requirements.txt
cryptography==3.4.8
Pillow==8.4.0
gunicorn==21.2.0
Flask==2.0.1
Werkzeug==2.0.3
